+++ language-server/Shared.AnZwDev.ALTools/CodeTransformations/SyntaxNodesGroupsTree.cs
@@ -6,6 +6,7 @@ using Microsoft.Dynamics.Nav.CodeAnalysis;
 using Microsoft.Dynamics.Nav.CodeAnalysis.Syntax;
 using AnZwDev.ALTools.Extensions;
 using AnZwDev.ALTools.ALSymbols.Internal;
+using System.Xml.Linq;
 
 namespace AnZwDev.ALTools.CodeTransformations
 {
@@ -21,7 +22,7 @@ namespace AnZwDev.ALTools.CodeTransformations
 
         #region Add nodes to the tree
 
-        public bool AddNodes(IEnumerable<T> nodesCollection)
+        public bool AddNodes(IEnumerable<T> nodesCollection, SyntaxTriviaList? endingTriviaList = null)
         {
             this.Root = new SyntaxNodesGroup<T>();
             SyntaxNodesGroup<T> group = this.Root;
@@ -34,6 +35,27 @@ namespace AnZwDev.ALTools.CodeTransformations
                     return false;
                 }
             }
+
+            if (endingTriviaList != null)
+            {
+                var hasGroups = false;
+                (group, hasGroups) = ProcessSyntaxTrivias(group, endingTriviaList.Value);
+                {
+                    if (group == null)
+                    {
+                        this.Root = null;
+                        return false;
+                    }
+                }
+            }
+
+            //something went wrong and we are not back at the top group (missing endregion directives)
+            if (group.ParentGroup != null)
+            {
+                this.Root = null;
+                return false;
+            }
+
             return true;
         }
 
@@ -87,8 +109,54 @@ namespace AnZwDev.ALTools.CodeTransformations
             group.SyntaxNodes.Add(node);
             return group;
         }
-        
-#endregion
+
+        public (SyntaxNodesGroup<T>, bool) ProcessSyntaxTrivias(SyntaxNodesGroup<T> group, SyntaxTriviaList syntaxTrivias)
+        {
+            bool hasGroups = false;
+
+            if ((syntaxTrivias != null) && (syntaxTrivias.Count > 0))
+            {
+                //collect regions
+                List<SyntaxTrivia> triviaCache = new List<SyntaxTrivia>();
+
+                foreach (SyntaxTrivia trivia in syntaxTrivias)
+                {
+                    triviaCache.Add(trivia);
+                    ConvertedSyntaxKind localTriviaKind = trivia.Kind.ConvertToLocalType();
+
+                    switch (localTriviaKind)
+                    {
+                        case ConvertedSyntaxKind.RegionDirectiveTrivia:
+                            SyntaxNodesGroup<T> childGroup = new SyntaxNodesGroup<T>();
+                            childGroup.LeadingTrivia = triviaCache;
+                            group.AddGroup(childGroup);
+                            group = childGroup;
+                            triviaCache = new List<SyntaxTrivia>();
+                            hasGroups = true;
+                            break;
+                        case ConvertedSyntaxKind.EndRegionDirectiveTrivia:
+                            group.TrailingTrivia = triviaCache;
+                            group = group.ParentGroup;
+                            if (group == null)
+                                return (null, hasGroups);
+                            triviaCache = new List<SyntaxTrivia>();
+                            hasGroups = true;
+                            break;
+                        default:
+#if BC
+                            //do not sort if code contains other directives
+                            if (trivia.IsDirective)
+                                return (null, hasGroups);
+#endif
+                            break;
+                    }
+                }
+            }
+
+            return (group, hasGroups);
+        }
+
+        #endregion
 
         public SyntaxList<T> CreateSyntaxList()
         {